+++ src/view/Home.vue
@@ -44,22 +44,33 @@
                     type="text"
                     class="form-control"
                     id="floatingInput"
-                    placeholder="Merk Motor"
+                    placeholder="Merk Mobil"
                     v-model="formMobil.name"
                     required
                   />
-                  <label for="floatingInput">Merk Motor</label>
+                  <label for="floatingInput">Merk Mobil</label>
                 </div>
                 <div class="form-floating mb-3">
                   <input
                     type="text"
                     class="form-control"
                     id="floatingInput"
-                    placeholder="Merk Motor"
+                    placeholder="Merk Mobil"
                     v-model="formMobil.tahun"
                     required
                   />
                   <label for="floatingInput">Tahun Buat</label>
+                </div>
+                <div class="form-floating mb-3">
+                  <input
+                    type="text"
+                    class="form-control"
+                    id="floatingInput"
+                    placeholder="Merk Mobil"
+                    v-model="formMobil.harga"
+                    required
+                  />
+                  <label for="floatingInput">Harga</label>
                 </div>
 
                 <button
@@ -97,8 +108,9 @@
         <thead>
           <tr style="text-align: center">
             <th>No.</th>
-            <th>Merk Motor</th>
+            <th>Merk Mobil</th>
             <th>Tahun Buatan</th>
+            <th>Harga</th>
             <th>Action</th>
           </tr>
         </thead>
@@ -112,6 +124,7 @@
             <td>{{ index + 1 }}</td>
             <td>{{ mobil.name }}</td>
             <td>{{ mobil.tahun }}</td>
+            <td>{{ mobil.harga }} jt</td>
             <td>
               <button
                 type="button"
@@ -126,6 +139,10 @@
               <button class="btn btn-danger" @click="delMobil(mobil)">
                 Delete
               </button>
+              ||
+              <button class="btn btn-warning" @click="buyMobil(mobil)">
+                Buy
+              </button>
             </td>
           </tr>
         </tbody>
@@ -137,6 +154,7 @@
 <script>
 import Navbar from "../components/Navbar.vue";
 import axios from "axios";
+import Swal from "sweetalert2";
 export default {
   name: "HomePage",
   components: {
@@ -148,10 +166,16 @@ export default {
         id: "",
         name: "",
         tahun: "",
+        harga: "",
+        type: "mobil",
       },
-      mobils: "",
+      mobils: [],
       updateSubmitMobil: false,
     };
+  },
+  props: {
+    cart: Array,
+    setCart: Function,
   },
   mounted() {
     if (!sessionStorage.getItem("USER_DATA")) {
@@ -175,6 +199,8 @@ export default {
         this.loadMobil();
         this.formMobil.name = "";
         this.formMobil.tahun = "";
+        this.formMobil.harga = "";
+        this.formMobil.type = "Mobil";
       });
       window.location.reload();
     },
@@ -184,6 +210,7 @@ export default {
       this.formMobil.id = mobil.id;
       this.formMobil.name = mobil.name;
       this.formMobil.tahun = mobil.tahun;
+      this.formMobil.harga = mobil.harga;
     },
     // Mobil update
     updateMobil(formMobil) {
@@ -197,6 +224,7 @@ export default {
           this.formMobil.id = "";
           this.formMobil.name = "";
           this.formMobil.tahun = "";
+          this.formMobil.harga = "";
           window.location.reload();
         })
         .catch((err) => {
@@ -205,15 +233,35 @@ export default {
     },
     // Delete Mobil
     delMobil(mobil) {
-      axios.delete("http://localhost:3000/mobils/" + mobil.id).then(() => {
-        this.loadMobil();
-        let index = this.mobils.indexOf();
-        this.mobils.splice(index, 1);
+      Swal.fire({
+        title: "Anda Yakin Menghapus Data ?",
+        showDenyButton: true,
+        icon: "question",
+        confirmButtonText: "Delete",
+        denyButtonText: "Batal",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          axios.delete("http://localhost:3000/mobils/" + mobil.id).then(() => {
+            this.loadMobil();
+            let index = this.mobils.indexOf();
+            this.mobils.splice(index, 1);
+          });
+        }
       });
     },
     // Close
     close() {
       window.location.reload();
+    },
+    buyMobil(mobil) {
+      if (this.cart.filter((x) => x.id === mobil.id && x.type === "mobil")[0])
+        return Swal.fire({
+          icon: "error",
+          title: "Oops...",
+          text: `Mobil ${mobil.name} sudah ada di cart`,
+        });
+      this.setCart([...this.cart, mobil]);
+      Swal.fire("Berhasil", `Berhasil membeli mobil ${mobil.name}`, "success");
     },
   },
 };
